Create python/decimal.{h,cc}
