+++ cpp/src/arrow/python/builtin_convert.cc
@@ -33,6 +33,7 @@
 #include "arrow/util/decimal.h"
 #include "arrow/util/logging.h"
 
+#include "arrow/python/decimal.h"
 #include "arrow/python/helpers.h"
 #include "arrow/python/numpy_convert.h"
 #include "arrow/python/util/datetime.h"
