+++ cpp/src/arrow/python/helpers.h
@@ -15,8 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#ifndef PYARROW_HELPERS_H
-#define PYARROW_HELPERS_H
+#ifndef ARROW_PYTHON_HELPERS_H
+#define ARROW_PYTHON_HELPERS_H
 
 #include "arrow/python/platform.h"
 
@@ -32,8 +32,6 @@
 #include "arrow/util/visibility.h"
 
 namespace arrow {
-
-class Decimal128;
 
 namespace py {
 
@@ -125,68 +123,8 @@ Status BuilderAppend(BinaryBuilder* builder, PyObject* obj, bool* is_full = null
 Status BuilderAppend(FixedSizeBinaryBuilder* builder, PyObject* obj,
                      bool* is_full = nullptr);
 
-//
-// Decimal helpers
-// XXX decimal.h?
-//
-
-// \brief Import
-Status ImportDecimalType(OwnedRef* decimal_type);
-
-// \brief Convert a Python Decimal object to a C++ string
-// \param[in] python_decimal A Python decimal.Decimal instance
-// \param[out] The string representation of the Python Decimal instance
-// \return The status of the operation
-Status PythonDecimalToString(PyObject* python_decimal, std::string* out);
-
-// \brief Convert a C++ std::string to a Python Decimal instance
-// \param[in] decimal_constructor The decimal type object
-// \param[in] decimal_string A decimal string
-// \return An instance of decimal.Decimal
-PyObject* DecimalFromString(PyObject* decimal_constructor,
-                            const std::string& decimal_string);
-
-// \brief Convert a Python decimal to an Arrow Decimal128 object
-// \param[in] python_decimal A Python decimal.Decimal instance
-// \param[in] arrow_type An instance of arrow::DecimalType
-// \param[out] out A pointer to a Decimal128
-// \return The status of the operation
-Status DecimalFromPythonDecimal(PyObject* python_decimal, const DecimalType& arrow_type,
-                                Decimal128* out);
-
-// \brief Check whether obj is an instance of Decimal
-bool PyDecimal_Check(PyObject* obj);
-
-// \brief Check whether obj is nan. This function will abort the program if the argument
-// is not a Decimal instance
-bool PyDecimal_ISNAN(PyObject* obj);
-
-// \brief Helper class to track and update the precision and scale of a decimal
-class DecimalMetadata {
- public:
-  DecimalMetadata();
-  DecimalMetadata(int32_t precision, int32_t scale);
-
-  // \brief Adjust the precision and scale of a decimal type given a new precision and a
-  // new scale \param[in] suggested_precision A candidate precision \param[in]
-  // suggested_scale A candidate scale \return The status of the operation
-  Status Update(int32_t suggested_precision, int32_t suggested_scale);
-
-  // \brief A convenient interface for updating the precision and scale based on a Python
-  // Decimal object \param object A Python Decimal object \return The status of the
-  // operation
-  Status Update(PyObject* object);
-
-  int32_t precision() const { return precision_; }
-  int32_t scale() const { return scale_; }
-
- private:
-  int32_t precision_;
-  int32_t scale_;
-};
-
 }  // namespace internal
 }  // namespace py
 }  // namespace arrow
 
-#endif  // PYARROW_HELPERS_H
+#endif  // ARROW_PYTHON_HELPERS_H
